[DataGrid] Move column headers virtualization to hook
