+++ docs/pages/api-docs/data-grid/data-grid-pro.json
@@ -243,8 +243,9 @@
       "columnHeaderDropZone",
       "columnHeaderTitle",
       "columnHeaderTitleContainer",
-      "columnHeaderWrapper",
-      "columnsContainer",
+      "columnHeaders",
+      "columnHeadersInner",
+      "columnHeadersInner--scrollable",
       "columnSeparator--resizable",
       "columnSeparator--resizing",
       "columnSeparator",
