+++ docs/pages/api-docs/data-grid/data-grid.json
@@ -215,8 +215,9 @@
       "columnHeaderDropZone",
       "columnHeaderTitle",
       "columnHeaderTitleContainer",
-      "columnHeaderWrapper",
-      "columnsContainer",
+      "columnHeaders",
+      "columnHeadersInner",
+      "columnHeadersInner--scrollable",
       "columnSeparator--resizable",
       "columnSeparator--resizing",
       "columnSeparator",
