+++ docs/translations/api-docs/data-grid/data-grid-pro-zh.json
@@ -216,13 +216,18 @@
       "description": "Styles applied to {{nodeName}}.",
       "nodeName": "the column header's title container element"
     },
-    "columnHeaderWrapper": {
+    "columnHeaders": {
       "description": "Styles applied to {{nodeName}}.",
-      "nodeName": "the column header's wrapper element"
+      "nodeName": "the column headers"
     },
-    "columnsContainer": {
+    "columnHeadersInner": {
       "description": "Styles applied to {{nodeName}}.",
-      "nodeName": "the outer columns container element"
+      "nodeName": "the column headers's inner element"
+    },
+    "columnHeadersInner--scrollable": {
+      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
+      "nodeName": "the column headers's inner element",
+      "conditions": "there is a horizontal scrollbar"
     },
     "columnSeparator--resizable": {
       "description": "Styles applied to {{nodeName}} if {{conditions}}.",
