+++ packages/grid/_modules_/grid/components/base/GridBody.tsx
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types';
 import { GridEvents } from '../../constants/eventsConstants';
 import { useGridApiContext } from '../../hooks/utils/useGridApiContext';
 import { ElementSize } from '../../models/elementSize';
-import { GridColumnsHeader } from '../columnHeaders/GridColumnHeaders';
-import { GridColumnsContainer } from '../containers/GridColumnsContainer';
 import { GridMainContainer } from '../containers/GridMainContainer';
 import { GridAutoSizer } from '../GridAutoSizer';
 import { GridOverlays } from './GridOverlays';
@@ -13,14 +11,17 @@ import { GridVirtualScroller } from '../GridVirtualScroller';
 import { useGridSelector } from '../../hooks/utils/useGridSelector';
 import { gridSelectionStateSelector } from '../../hooks/features/selection/gridSelectionSelector';
 import { gridDensityHeaderHeightSelector } from '../../hooks/features/density/densitySelector';
-import { GridScrollArea } from '../GridScrollArea';
 
 interface GridBodyProps {
   children?: React.ReactNode;
+  ColumnHeadersComponent: React.JSXElementConstructor<{
+    ref: React.Ref<HTMLDivElement>;
+    innerRef: React.Ref<HTMLDivElement>;
+  }>;
 }
 
 function GridBody(props: GridBodyProps) {
-  const { children } = props;
+  const { children, ColumnHeadersComponent } = props;
   const apiRef = useGridApiContext();
   const rootProps = useGridRootProps();
   const selection = useGridSelector(apiRef, gridSelectionStateSelector);
@@ -81,11 +82,7 @@ function GridBody(props: GridBodyProps) {
   return (
     <GridMainContainer>
       <GridOverlays />
-      <GridScrollArea scrollDirection="left" />
-      <GridColumnsContainer ref={columnsContainerRef}>
-        <GridColumnsHeader ref={columnsHeaderRef} />
-      </GridColumnsContainer>
-      <GridScrollArea scrollDirection="right" />
+      <ColumnHeadersComponent ref={columnsContainerRef} innerRef={columnsHeaderRef} />
       <GridAutoSizer
         nonce={rootProps.nonce}
         disableHeight={rootProps.autoHeight}
@@ -121,6 +118,7 @@ GridBody.propTypes = {
   // | To update them edit the TypeScript types and run "yarn proptypes"  |
   // ----------------------------------------------------------------------
   children: PropTypes.node,
+  ColumnHeadersComponent: PropTypes.elementType.isRequired,
 } as any;
 
 export { GridBody };
