+++ packages/grid/_modules_/grid/components/columnHeaders/index.ts
@@ -2,5 +2,3 @@ export * from './GridColumnHeaderItem';
 export * from './GridColumnHeaderSeparator';
 export * from './GridColumnHeaderSortIcon';
 export * from './GridColumnHeaderTitle';
-export * from './GridColumnHeaders';
-export * from './GridColumnHeadersItemCollection';
