+++ packages/grid/_modules_/grid/components/containers/index.ts
@@ -1,5 +1,4 @@
 export * from './GridRoot';
-export * from './GridColumnsContainer';
 export * from './GridFooterContainer';
 export * from './GridOverlay';
 export * from './GridToolbarContainer';
