+++ packages/grid/x-grid/src/DataGridPro.tsx
@@ -15,6 +15,7 @@ import { useDataGridProComponent } from './useDataGridProComponent';
 import { Watermark } from '../../_modules_/grid/components/Watermark';
 import { DataGridProProps } from './DataGridProProps';
 import { useDataGridProProps } from './useDataGridProProps';
+import { DataGridProColumnHeaders } from './DataGridProColumnHeaders';
 
 // This is the package release date. Each package version should update this const
 // automatically when a new version is published on npm.
@@ -41,7 +42,7 @@ const DataGridProRaw = React.forwardRef<HTMLDivElement, DataGridProProps>(functi
       <GridRoot ref={ref}>
         <GridErrorHandler>
           <GridHeaderPlaceholder />
-          <GridBody>
+          <GridBody ColumnHeadersComponent={DataGridProColumnHeaders}>
             <Watermark />
           </GridBody>
           <GridFooterPlaceholder />
