+++ scripts/exportsSnapshot.json
@@ -6,7 +6,6 @@
   { "name": "GridColumnHeaderItem", "kind": "Namespace" },
   { "name": "GridColumnHeaderMenu", "kind": "Namespace" },
   { "name": "GridColumnHeaderTitle", "kind": "Namespace" },
-  { "name": "GridColumnHeadersItemCollection", "kind": "Namespace" },
   { "name": "GridColumnsMenuItem", "kind": "Namespace" },
   { "name": "GridEditInputCell", "kind": "Namespace" },
   { "name": "GridEditSingleSelectCell", "kind": "Namespace" },
@@ -47,7 +46,6 @@
   { "name": "GridColumnHeaderSeparatorProps", "kind": "Interface" },
   { "name": "GridColumnHeaderSortIconProps", "kind": "Interface" },
   { "name": "GridColumnHeaderTitleProps", "kind": "Interface" },
-  { "name": "GridColumnHeadersItemCollectionProps", "kind": "Interface" },
   { "name": "GridColumnMenuApi", "kind": "Interface" },
   { "name": "GridColumnMenuProps", "kind": "Interface" },
   { "name": "GridColumnMenuState", "kind": "Interface" },
@@ -227,8 +225,6 @@
   { "name": "GridColumnIcon", "kind": "Variable" },
   { "name": "GridColumnMenu", "kind": "Variable" },
   { "name": "GridColumnMenuContainer", "kind": "Variable" },
-  { "name": "GridColumnsContainer", "kind": "Variable" },
-  { "name": "GridColumnsHeader", "kind": "Variable" },
   { "name": "GridColumnsMenuItem", "kind": "Variable" },
   { "name": "GridDragIcon", "kind": "Variable" },
   { "name": "GridFeatureModeConstant", "kind": "Variable" },
@@ -345,7 +341,6 @@
   { "name": "GridColumnHeaderItem", "kind": "Function" },
   { "name": "GridColumnHeaderMenu", "kind": "Function" },
   { "name": "GridColumnHeaderTitle", "kind": "Function" },
-  { "name": "GridColumnHeadersItemCollection", "kind": "Function" },
   { "name": "GridColumnsPanel", "kind": "Function" },
   { "name": "GridEditInputCell", "kind": "Function" },
   { "name": "GridEditSingleSelectCell", "kind": "Function" },
